Python and Resource files upload
Uploaded the resources (for e.g. images used inside the python code), the python code tryfacerec.py to recognize or authorize face and send data to arduino which can be further processed in arduino, the python code testdatacreation,py which is for creating the dataset for capturing the frames and storing it inside a folder called "dataset", then at last we have encode_generator.py which is to create encoding with the captured frames from the dataset.